+++ modules/ensemble/lib/ensemble_app.dart
@@ -30,6 +30,7 @@ import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:flutter_localized_locales/flutter_localized_locales.dart';
 import 'package:workmanager/workmanager.dart';
 import 'package:yaml/yaml.dart';
+import 'package:connectivity_plus/connectivity_plus.dart';
 
 const String backgroundUploadTask = 'backgroundUploadTask';
 const String backgroundBluetoothSubscribeTask = 'backgroundBluetoothSubscribeTask';
@@ -141,11 +142,22 @@ class EnsembleAppState extends State<EnsembleApp> with WidgetsBindingObserver {
   // the entire App to reload (e.g. change Locale at runtime)
   Key? appKey;
 
+  bool _hasInternet = true;
+  late final StreamSubscription<List<ConnectivityResult>>
+      _connectivitySubscription;
+
   @override
   void initState() {
     super.initState();
     WidgetsBinding.instance.addObserver(this);
     config = initApp();
+
+    // Initialize connectivity listener.
+    _updateConnectivity();
+    _connectivitySubscription = Connectivity()
+        .onConnectivityChanged
+        .listen((_) => _updateConnectivity());
+
     // Initialize native features.
     if (!kIsWeb) {
       Workmanager().initialize(callbackDispatcher, isInDebugMode: false);
@@ -173,6 +185,23 @@ class EnsembleAppState extends State<EnsembleApp> with WidgetsBindingObserver {
     }
   }
 
+  /// Check the device’s connectivity and update the state.
+  Future<void> _updateConnectivity() async {
+    final result = await Connectivity().checkConnectivity();
+    final hasInternetNow = result.any((r) => r != ConnectivityResult.none);
+
+    // If connectivity has been restored, reinitialize the app
+    if (!_hasInternet && hasInternetNow) {
+      setState(() {
+        config = initApp();
+      });
+    }
+
+    setState(() {
+      _hasInternet = hasInternetNow;
+    });
+  }
+
   @override
   void didChangeAppLifecycleState(AppLifecycleState state) async {
     super.didChangeAppLifecycleState(state);
@@ -192,6 +221,7 @@ class EnsembleAppState extends State<EnsembleApp> with WidgetsBindingObserver {
 
   @override
   void dispose() {
+    _connectivitySubscription.cancel();
     WidgetsBinding.instance.removeObserver(this);
     super.dispose();
   }