+++ modules/ensemble/lib/page_model.dart
@@ -353,6 +353,7 @@ class SinglePageModel extends PageModel with HasStyles {
     WidgetModel? titleWidget;
     String? titleText = legacyTitle;
     WidgetModel? background;
+    WidgetModel? leadingWidget;
     Map<String, dynamic>? styles;
     List<String>? classList;
 
@@ -372,6 +373,11 @@ class SinglePageModel extends PageModel with HasStyles {
         }
       }
 
+      if (headerData['leadingWidget'] != null) {
+        leadingWidget = ViewUtil.buildModel(
+            headerData['leadingWidget'], customViewDefinitions);
+      }
+
       if (headerData['flexibleBackground'] != null) {
         background = ViewUtil.buildModel(
             headerData['flexibleBackground'], customViewDefinitions);
@@ -386,11 +392,13 @@ class SinglePageModel extends PageModel with HasStyles {
         titleText != null ||
         background != null ||
         styles != null ||
+        leadingWidget != null ||
         classList != null) {
       headerModel = HeaderModel(
           titleText: titleText,
           titleWidget: titleWidget,
           flexibleBackground: background,
+          leadingWidget: leadingWidget,  
           inlineStyles: styles,
           classList: classList);
     }
@@ -533,6 +541,7 @@ class HeaderModel extends Object with HasStyles {
       {this.titleText,
       this.titleWidget,
       this.flexibleBackground,
+      this.leadingWidget,
       inlineStyles,
       classList}) {
     this.inlineStyles = inlineStyles;
@@ -542,6 +551,7 @@ class HeaderModel extends Object with HasStyles {
   // header title can be text or a widget
   String? titleText;
   WidgetModel? titleWidget;
+  WidgetModel? leadingWidget;
 
   WidgetModel? flexibleBackground;
 }