+++ modules/ensemble/lib/util/utils.dart
@@ -3,8 +3,12 @@ import 'dart:math';
 import 'dart:ui';
 import 'package:ensemble/ensemble.dart';
 import 'package:ensemble/ensemble_app.dart';
+import 'package:ensemble/framework/action.dart';
+import 'package:ensemble/framework/ensemble_config_service.dart';
 import 'package:ensemble/framework/stub/location_manager.dart';
 import 'package:ensemble/framework/theme/theme_manager.dart';
+import 'package:ensemble/screen_controller.dart';
+import 'package:ensemble/widget/helpers/tooltip_composite.dart';
 import 'package:ensemble_ts_interpreter/invokables/UserLocale.dart';
 import 'package:path/path.dart' as p;
 
@@ -488,6 +492,81 @@ class Utils {
     return null;
   }
 
+  // Creates tooltip composite from inputs
+  static TooltipStyleComposite? getTooltipStyleComposite(
+      ChangeNotifier controller, dynamic inputs) {
+    if (inputs is Map) {
+      return TooltipStyleComposite(controller, inputs: inputs);
+    }
+    return null;
+  }
+
+  /// Creates tooltip widget with configured styles and behavior
+  static Widget getTooltipWidget(
+  BuildContext context,
+  Widget child,
+  Map<String, dynamic>? tooltipData,
+  ChangeNotifier controller
+) {
+  if (tooltipData == null) return child;
+  
+  final tooltip = TooltipData.from(tooltipData, controller);
+    if (tooltip == null) return child;
+
+    final tooltipKey = GlobalKey();
+    // Start with the original child
+    Widget tooltipChild = child;
+
+    if (kIsWeb && tooltip.styles?.triggerMode == null) {
+      tooltipChild = MouseRegion(
+        onEnter: (_) {
+          final dynamic tooltip = tooltipKey.currentState;
+          tooltip?.ensureTooltipVisible();
+        },
+        onExit: (_) {
+          final dynamic tooltip = tooltipKey.currentState;
+          tooltip?.deactivate();
+        },
+        child: tooltipChild,
+      );
+    }
+
+    return Tooltip(
+      key: tooltipKey,
+      message: tooltip.message,
+      textStyle: tooltip.styles?.textStyle,
+      padding: tooltip.styles?.padding,
+      margin: tooltip.styles?.margin,
+      verticalOffset: tooltip.styles?.verticalOffset,
+      preferBelow: tooltip.styles?.preferBelow,
+      waitDuration:
+          tooltip.styles?.waitDuration ?? const Duration(milliseconds: 0),
+      showDuration:
+          tooltip.styles?.showDuration ?? const Duration(milliseconds: 1500),
+      triggerMode: tooltip.styles?.triggerMode ?? TooltipTriggerMode.tap,
+      enableFeedback: true,
+      decoration: BoxDecoration(
+        color: tooltip.styles?.backgroundColor ?? Colors.grey[700],
+        borderRadius: tooltip.styles?.borderRadius,
+        border: (tooltip.styles?.borderColor != null ||
+                tooltip.styles?.borderWidth != null)
+            ? Border.all(
+                color: tooltip.styles?.borderColor ??
+                    ThemeManager().getBorderColor(context),
+                width: (tooltip.styles?.borderWidth ??
+                        ThemeManager().getBorderThickness(context))
+                    .toDouble(),
+              )
+            : null,
+      ),
+      onTriggered: tooltip.onTriggered != null
+          ? () =>
+              ScreenController().executeAction(context, tooltip.onTriggered!)
+          : null,
+      child: tooltipChild,
+    );
+  }
+
   static BoxShadowComposite? getBoxShadowComposite(
       ChangeNotifier widgetController, dynamic inputs) {
     if (inputs is Map) {
@@ -615,13 +694,6 @@ class Utils {
     return textAlign;
   }
 
-  static double? getValidOpacity(double opacity) {
-    if (opacity < 0 || opacity > 1) {
-      return 1;
-    } else {
-      return opacity;
-    }
-  }
 
   static Curve? getCurve(String? curveType) {
     Curve? curve;
@@ -996,10 +1068,21 @@ static BoxDecoration? getBoxDecoration(dynamic style) {
     return strings[0];
   }
 
-  /// prefix the asset with the root directory (i.e. ensemble/assets/), plus
+  /// prefix the asset with the app root directory (i.e. ensemble/apps/<app-name>/assets/), plus
   /// stripping any unnecessary query params (e.g. anything after the first ?)
   static String getLocalAssetFullPath(String asset) {
-    return 'ensemble/assets/${stripQueryParamsFromAsset(asset)}';
+    try {
+      String provider = EnsembleConfigService.config["definitions"]?['from'];
+      if (provider == 'local') {
+        String path =
+            EnsembleConfigService.config["definitions"]?['local']?["path"];
+        return '${path}/assets/${stripQueryParamsFromAsset(asset)}';
+      } else {
+        return 'ensemble/assets/${stripQueryParamsFromAsset(asset)}';
+      }
+    } catch (e) {
+      return 'ensemble/assets/${stripQueryParamsFromAsset(asset)}';
+    }
   }
 
   static bool isMemoryPath(String path) {