+++ modules/ensemble/lib/widget/widget_registry.dart
@@ -60,7 +60,6 @@ import 'package:ensemble/widget/stub_widgets.dart';
 import 'package:ensemble/widget/switch.dart';
 import 'package:ensemble/widget/text.dart';
 import 'package:ensemble/widget/toggle_button.dart';
-import 'package:ensemble/widget/tooltip.dart';
 import 'package:ensemble/widget/video.dart';
 import 'package:ensemble/widget/slidable.dart';
 import 'package:ensemble/widget/accordion.dart';
@@ -140,7 +139,6 @@ class WidgetRegistry {
         PopupMenu.type: () => PopupMenu(),
         EnsembleCalendar.type: () => EnsembleCalendar(),
         Countdown.type: () => Countdown(),
-        ToolTip.type: () => ToolTip(),
         EnsembleAccordion.type: () => EnsembleAccordion(),
 
         // form fields